Fix Recovery OS Rust build: resolve all 3 type errors
- recovery_ota.rs: esp_ota_write expects usize, not u32 for length
- recovery_web.rs: add explicit anyhow::Result<()> return types on
  all HTTP handler closures (compiler can't infer E: Debug)
- main.rs: ap.auth_method is Option<AuthMethod>, wrap in Some()
- Remove unused imports (std::net::Ipv4Addr, std::ffi::CString,
  std::io::Read)

Recovery OS now compiles cleanly with zero warnings.

+++ recovery/src/recovery_ota.rs
@@ -4,8 +4,6 @@
 use esp_idf_svc::http::client::{Configuration as HttpConfig, EspHttpConnection};
 use esp_idf_sys::*;
 use log::*;
-use std::ffi::CString;
-use std::io::Read;
 
 const SD_FIRMWARE_PATH: &str = "/sdcard/update/thistle_os.bin";
 const MAX_FIRMWARE_SIZE: usize = 4 * 1024 * 1024; // 4MB
@@ -122,7 +120,7 @@ fn flash_to_ota1(data: &[u8]) -> anyhow::Result<()> {
         let mut written = 0;
         while written < total {
             let end = std::cmp::min(written + chunk_size, total);
-            let ret = esp_ota_write(handle, data[written..end].as_ptr() as *const _, (end - written) as u32);
+            let ret = esp_ota_write(handle, data[written..end].as_ptr() as *const _, end - written);
             if ret != ESP_OK as i32 {
                 esp_ota_abort(handle);
                 anyhow::bail!("esp_ota_write failed at offset {}: {}", written, ret);
